[IMP] progressbar: set textual progress as a fractional number
Right now, the widget progressbar shows the textual progress
as either a percentage or a fractional number. The user has
no way of controlling when the progress is a fractional number,
because by default, if the maximum value is 100, the textual
progress is shown as a percentage. This has the problem of
showing the progress sometimes as a percentage and sometimes
as a fraction, for example in these cases in which the maximum
value is dynamic and can change over time.

With the proposed improvement, a new option has been added to
the widget, `show_fractional`, that if set to true, displays the
textual progress always using the fractional representation.